omit empty coupon fields in XML
This makes optional arguments to CreateCoupon optional. Recurly only
requires `coupon_code`, `name`, and `discount_type`.

When these were not omitempty, we received errors from Recurly's API
(`the provided xml was invalid`).

Additionally, the boolean arguments defaulted to `false` when not
explicitly set to `true`, which was a surprise as Recurly assumes some
of them default to `true` (e.g. `applies_to_all_plans`).

After this change, Recurly accepts the XML payload and implicit defaults are
as expected.

After this change, the XML is accepted by Recurly with expected defaults.